[Nexters/covid-letter-web#45] 🚚 type file rename
