+++ src/pages/api/covid/stats.ts
@@ -1,7 +1,7 @@
 import {createResponse} from '$utils/fetcher/withAxios'
 import {Response} from '$types/response'
 import {NextApiRequest, NextApiResponse} from 'next'
-import {CovidStats} from '$types/response/analyze'
+import {CovidStats} from '$types/response/stat'
 
 const routes = async (_req: NextApiRequest, res: NextApiResponse<Response<CovidStats>>) => {
     res.setHeader('Set-Cookie', [`letterLogin=; path=/; expires=-1`, `googleLogin=; path=/; expires=-1`])
