+++ src/pages/api/letter/stats.ts
@@ -1,7 +1,7 @@
 import {createResponse} from '$utils/fetcher/withAxios'
 import {Response} from '$types/response'
 import {NextApiRequest, NextApiResponse} from 'next'
-import {LetterStats} from '$types/response/analyze'
+import {LetterStats} from '$types/response/stat'
 
 const routes = async (_req: NextApiRequest, res: NextApiResponse<Response<LetterStats>>) => {
     res.setHeader('Set-Cookie', [`letterLogin=; path=/; expires=-1`, `googleLogin=; path=/; expires=-1`])
